backend/requirements.txt
flask
flask-cors
openai
gspread
oauth2client
python-dotenv
gunicorn
pytz